requirements.txt
playwright==1.44.0
pandas==2.2.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
openpyxl==3.1.2
